HTBHF-950 refactor test data factories to use consistent naming convention
