+++ api/src/test/java/uk/gov/dhsc/htbhf/hmrc/converter/EligibilityRequestToHMRCEligibilityRequestTest.java
@@ -5,7 +5,7 @@ import uk.gov.dhsc.htbhf.hmrc.model.EligibilityRequest;
 import uk.gov.dhsc.htbhf.hmrc.model.HMRCEligibilityRequest;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static uk.gov.dhsc.htbhf.hmrc.testhelper.EligibilityRequestTestFactory.aValidEligibilityRequest;
+import static uk.gov.dhsc.htbhf.hmrc.testhelper.EligibilityRequestTestDataFactory.aValidEligibilityRequest;
 
 public class EligibilityRequestToHMRCEligibilityRequestTest {
     private EligibilityRequestToHMRCEligibilityRequest converter = new EligibilityRequestToHMRCEligibilityRequest();
