+++ api/src/test/java/uk/gov/dhsc/htbhf/hmrc/model/PersonDTOTest.java
@@ -11,7 +11,7 @@ import javax.validation.ConstraintViolation;
 
 import static uk.gov.dhsc.htbhf.assertions.ConstraintViolationAssert.assertThat;
 import static uk.gov.dhsc.htbhf.hmrc.testhelper.AddressDTOTestDataFactory.anAddressWithAddressLine1;
-import static uk.gov.dhsc.htbhf.hmrc.testhelper.PersonDTOTestFactory.*;
+import static uk.gov.dhsc.htbhf.hmrc.testhelper.PersonDTOTestDataFactory.*;
 
 class PersonDTOTest extends AbstractValidationTest {
